+++ Remote2D/src/com/remote/remote2d/editor/GuiCreateSpriteSheet.java
@@ -101,7 +101,7 @@ public class GuiCreateSpriteSheet extends GuiMenu {
 		framesY.text = (int)anim.getFrames().y+"";
 		paddingX.text = (int)anim.getPadding().x+"";
 		paddingY.text = (int)anim.getPadding().y+"";
-		frameLength.text = (int)anim.getFramelength()+"";
+		frameLength.text = anim.getFramelength()+"";
 		animSave.text = anim.getPath();
 	}
